restructured according to REPO_STRUCTURE (#40)
