+++ code_soup/ch5/__init__.py
@@ -1,2 +1,1 @@
-from code_soup.ch5.datasets import MnistDataset
 from code_soup.ch5.models import GAN
