--- code_soup/ch5/datasets/__init__.py
@@ -1,1 +1,0 @@
-from code_soup.ch5.datasets.mnist import MnistDataset
