+++ code_soup/common/vision/datasets/__init__.py
@@ -1,0 +1,1 @@
+from code_soup.common.vision.datasets.mnist import MnistDataset
