+++ code_soup/common/vision/datasets/readme.md
@@ -1,0 +1,6 @@
+# Common Datasets for vision
+
+List of implemented datasets
+---
+
+1. MNIST
