+++ code_soup/common/vision/models/readme.md
@@ -1,0 +1,4 @@
+# Common Models for Vision
+
+List of implemented models
+---
