+++ code_soup/common/vision/readme.md
@@ -1,0 +1,1 @@
+# Common Utility Modules for Vision
