+++ tests/test_ch5/__init__.py
@@ -1,4 +1,3 @@
-from tests.test_ch5.test_datasets import TestMnistDataset
 from tests.test_ch5.test_models import (
     TestDiscriminatorModel,
     TestGANModel,
