--- tests/test_ch5/test_datasets/__init__.py
@@ -1,1 +1,0 @@
-from tests.test_ch5.test_datasets.mnist_test import TestMnistDataset
