+++ tests/test_common/test_vision/test_datasets/__init__.py
@@ -1,0 +1,1 @@
+from tests.test_common.test_vision.test_datasets.mnist_test import TestMnistDataset
